Set names for density particles in native RMF
The PMI2 branch of the tutorial was failing because
when trying to calculate the accuracy of the models
RMF insists on particles being named identically in
the models and the known-good (native) structure,
which is unlikely given that density particles got
default (semi-random) particle names. Set nice names
to be consistent with those set by salilab/pmi@fceb9b8f.
This together with salilab/imp@119f96a *should*
resolve salilab/pmi#242 - will close that issue once
the Travis build completes successfully.