Put components in topology file order
Put components in the same order in the RMF file as
in the topology file. Previously they were in essentially
random order (the order that Python 2 emitted keys
from the molecules dict). This caused usage of this
RMF file to fail under Python 3, which emits keys in
a different random order. Keys in sorted order should
match behavior of PMI as of salilab/pmi@0b470f9.