Update modeling script to work with PMI2
Update the modeling script, keeping it as close
to the original as possible, to work with PMI2
in both the latest IMP stable release (2.9.0) and
IMP-develop. This incorporates changes from the
latest version of the tutorial (in the develop
branch) but does not use the new analysis yet,
as it is not complete or present in the stable
release. This should be considered an interim
solution so that we can safely deprecate PMI1.